Refactor chat-logic and renaming
