+++ src/features/Chat/index.tsx
@@ -1,9 +1,3 @@
-import { useGetContactsQuery, useGetMessagesQuery } from './chatPageApi';
-import { skipToken } from '@reduxjs/toolkit/dist/query';
-import { useAppSelector } from '@/store';
-import { selectUserId } from '../Authentication/userSlice';
-import { selectCurrentPollingParams } from './chatSlice';
-
 import { useMobileMode } from '@/hooks/useMobileMode';
 
 import styled, { css } from 'styled-components';
@@ -13,29 +7,20 @@ import {
   CONTENT_WIDTH,
   OUTER_VERTICAL_MARGIN,
 } from '@/components/variables';
-import ChatMenu from './components/Menu';
-import ChatWindow from './components/Window';
+import Menu from './components/Menu';
+import Window from './components/Window';
 import PageWithTransition from '@/components/PageWithTransition';
 
 const Chat = () => {
   const isMobile = useMobileMode();
 
-  const userId = useAppSelector(selectUserId);
-  const params = useAppSelector(selectCurrentPollingParams);
-  const messageParams = !!params && !!userId ? { params, userId } : skipToken;
-
-  useGetContactsQuery(userId ?? skipToken);
-  useGetMessagesQuery(messageParams, {
-    pollingInterval: 5000,
-  });
-
   return (
     <PageWithTransition>
       <PageContainer isMobile={isMobile}>
-        <ChatContainer>
-          <ChatMenu />
-          <ChatWindow />
-        </ChatContainer>
+        <Container>
+          <Menu />
+          <Window />
+        </Container>
       </PageContainer>
     </PageWithTransition>
   );
@@ -59,7 +44,7 @@ const PageContainer = styled.div<{ isMobile: boolean }>`
   }
 `;
 
-const ChatContainer = styled.div`
+const Container = styled.div`
   display: flex;
   gap: 22px;
   height: ${CONTENT_HEIGHT};
